base class turned to abstract ParrotType enum and create_parrot method moved to parrot_creation.py
